Rename Button to NotificationButton
