+++ src/provider/view/components/NotificationButton/NotificationButton.scss
